changed fader's default value for the native effect plugins
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@734 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/plugins/chorus/ChorusPlug.h
@@ -49,10 +49,10 @@ class		ChorusPlugin: public Plugin
   Chorus	*chorus2;
 
 
-  double	BaseLength;
-  double	ModDepth;
-  double	Frequency;
-  double	EffectMix;
+  float		BaseLength;
+  float		ModDepth;
+  float		Frequency;
+  float		EffectMix;
 
  protected:
   bool		Bypass;
